Added parsing of 'throws (Type)' syntax.
New diagnostic if the parsing fails. Logic to parse an optional
parenthesized type identifier after the 'throws' keyword.

This will not compile the standard library right now because it cannot
parse the pattern

  `func f(_ body: (T) throws -> U) rethrows -> U {return try body(t)}`

That pattern is really common in the standard library for things like
`withUnsafeBufferPointer`. I'm not sure what the problem is right now there.